requirements.txt
python-telegram-bot==21.5
openai==1.51.0
aiosqlite==0.20.0
aiohttp==3.10.5
aiofiles==24.1.0
pydantic==2.9.2
jinja2==3.1.4
python-dotenv==1.0.1
